requirements.txt
# ── Streamlit 앱 ──────────────────────────────────────────────
streamlit>=1.32.0

# ── 웹 스크래핑 ───────────────────────────────────────────────
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
readability-lxml>=0.8.1
trafilatura>=1.6.0

# ── RSS 수집 ──────────────────────────────────────────────────
feedparser>=6.0.0

# ── 데이터 분석·시각화 ───────────────────────────────────────
pandas>=2.0.0
plotly>=5.18.0

# ── LLM (Groq) ───────────────────────────────────────────────
groq>=0.4.0

# ── 환경변수 ──────────────────────────────────────────────────
python-dotenv>=1.0.0

# ── Google Sheets 피드백 저장 ──────────────────────────────
gspread>=5.12.0
google-auth>=2.20.0

# ── 테스트 ────────────────────────────────────────────────────
pytest>=7.0.0
